Extend base asset account, slightly rework tests
Add a support for pyth oracles and in place prices in BaseAssetInfo.
We optionally store them even if they are not used in the risk engine.
This info can be used for better matching of base asset with external
protocols. It also allows for a price extractor in the risk engine to
implement a more complex logic in the future.
Reserved zeroed space is added to BaseAssetInfo and
MintInfo for easier backward-compatible extentions.
In tests, on top of adding more tests for new functionality, assetToken
is renamed ot btcToken and additionalAssetToken to solToken.

STORAGE CHANGE: BaseAssetInfo and MintInfo accounts are extended in
a not backward-compatible way

INSTRUCTION CHANGE: add_base_asset and change_base_asset_parameters

TASK: PRO-79 and PRO-54